+++ trie.h
@@ -1,17 +1,17 @@
 #ifndef TRIE_H
 #define TRIE_H
 
-#include <stddef.h>
+#include <sys/types.h>
 
 typedef struct TrieNode {
     char ch;
-    int iCmd;
+    ssize_t iCmd;
 
     // For characters branching off the same node
-    int nextSibling;
+    ssize_t nextSibling;
 
     // For characters branching off of this one
-    int firstChild;
+    ssize_t firstChild;
 } TRIE_NODE;
 
 typedef struct Trie {
@@ -22,9 +22,9 @@ typedef struct Trie {
 
 TRIE trieGetNew(void);
 void triePrint(const TRIE* const trie);
-void trieAddString(TRIE* const trie, const int iCmd, const char str[const]);
+void trieAddString(TRIE* const trie, const ssize_t iCmd, const char str[const]);
 size_t trieAutocomplete(const TRIE* const trie, const size_t buflen, char usrBuf[const buflen]);
-int trieGetCmdIndex(const TRIE* const trie, const size_t buflen, char usrBuf[const buflen]);
+ssize_t trieGetCmdIndex(const TRIE* const trie, const size_t buflen, char usrBuf[const buflen]);
 int trieDelete(TRIE* const trie);
 
 #endif
